Initial commit of GLI-based cmd-line tool
This is aiming to be a command-line tool which has feature-parity with
both the Ant build and the Node (grunt) build.

It will be packaged up as a Ruby gem, which is far easier to consume for
many users than the current merge/pull in place for the ant build.